Remove MbedTLS TLS implementation
With the new OpenSSL implementation, the deprecated MbedTLS version can
be removed. This only removes the TLS part, not Crypto nor the
tls_client or any other use of MbedTLS directly in the code base.
