+++ src/tls/client.h
@@ -2,8 +2,21 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#ifdef TLS_PROVIDER_IS_MBEDTLS
-#  include "mbedtls/client.h"
-#else
-#  include "openssl/client.h"
-#endif
+#include "context.h"
+
+namespace tls
+{
+  class Client : public Context
+  {
+  private:
+    std::shared_ptr<Cert> cert;
+
+  public:
+    Client(std::shared_ptr<Cert> cert_, bool dtls = false) :
+      Context(true, dtls),
+      cert(cert_)
+    {
+      cert->use(ssl, cfg);
+    }
+  };
+}
