+++ src/tls/server.h
@@ -2,8 +2,21 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#ifdef TLS_PROVIDER_IS_MBEDTLS
-#  include "mbedtls/server.h"
-#else
-#  include "openssl/server.h"
-#endif
+#include "context.h"
+
+namespace tls
+{
+  class Server : public Context
+  {
+  private:
+    std::shared_ptr<Cert> cert;
+
+  public:
+    Server(std::shared_ptr<Cert> cert_, bool dtls = false) :
+      Context(false, dtls),
+      cert(cert_)
+    {
+      cert->use(ssl, cfg);
+    }
+  };
+}
